Rename: auth.parser, auth.validation. auth.js 폴더로 묶음
